Make service file more manageable
