+++ setup.cfg
@@ -1,3 +1,4 @@
 [flake8]
 exclude = .git,*migrations*, venv, aidants_connect/settings.py
 max-line-length = 88
+ignore = E203
